admin 통계 refactoring
