+++ backend/yigil-admin/src/main/java/kr/co/yigil/stats/domain/StatsServiceImpl.java
@@ -1,18 +1,41 @@
 package kr.co.yigil.stats.domain;
 
+import kr.co.yigil.favor.domain.DailyFavorCount;
+import kr.co.yigil.favor.domain.DailyFavorCountReader;
+import kr.co.yigil.region.domain.DailyRegion;
+import kr.co.yigil.travel.TravelType;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import java.time.LocalDate;
 import java.util.List;
-import kr.co.yigil.region.domain.DailyRegion;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class StatsServiceImpl implements StatsService {
     private final StatsReader statsReader;
+    private final DailyFavorCountReader dailyFavorCountReader;
 
     @Override
     public List<DailyRegion> getRegionStats(LocalDate startDate, LocalDate endDate) {
         return statsReader.getRegionStats(startDate, endDate);
+    }
+
+    @Transactional
+    @Override
+    public StaticInfo.DailyFavorsInfo getDailyFavors(LocalDate startDate, LocalDate endDate, TravelType travelType, Pageable pageable) {
+        Page<DailyFavorCount> dailyFavorCountPageable = dailyFavorCountReader.readDailyFavorCountBetween(startDate, endDate, travelType, pageable);
+        return new StaticInfo.DailyFavorsInfo(dailyFavorCountPageable);
+    }
+
+    @Transactional
+    @Override
+    public StaticInfo.DailyFavorsInfo getTopDailyFavors(LocalDate startDate, LocalDate endDate, TravelType travelType, Integer limit) {
+        List<DailyFavorCount> topDailyFavorCount = dailyFavorCountReader.getTopDailyFavorCount(startDate, endDate, travelType, limit);
+
+        return new StaticInfo.DailyFavorsInfo(topDailyFavorCount);
     }
 }
