+++ backend/yigil-admin/src/main/java/kr/co/yigil/stats/interfaces/dto/mapper/StatsMapper.java
@@ -1,12 +1,14 @@
 package kr.co.yigil.stats.interfaces.dto.mapper;
 
-import java.util.List;
 import kr.co.yigil.region.domain.DailyRegion;
-import kr.co.yigil.stats.interfaces.dto.response.RegionStatsResponse;
+import kr.co.yigil.stats.domain.StaticInfo;
+import kr.co.yigil.stats.interfaces.dto.StatisticsDto;
 import org.mapstruct.InjectionStrategy;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.ReportingPolicy;
+
+import java.util.List;
 
 @Mapper(
         componentModel = "spring",
@@ -15,10 +17,15 @@ import org.mapstruct.ReportingPolicy;
 )
 public interface StatsMapper {
 
-    default RegionStatsResponse toResponse(List<DailyRegion> dailyRegions) {
-        return new RegionStatsResponse(dailyRegions.stream().map(this::toDto).toList());
+    default StatisticsDto.RegionStatsResponse toResponse(List<DailyRegion> dailyRegions) {
+        return new StatisticsDto.RegionStatsResponse(dailyRegions.stream().map(this::toDto).toList());
     }
 
     @Mapping(target = "region", expression = "java(dailyRegion.getRegion().getName())")
-    RegionStatsResponse.RegionStatsInfo toDto(DailyRegion dailyRegion);
+    StatisticsDto.RegionStatsInfo toDto(DailyRegion dailyRegion);
+
+    @Mapping(target = "dailyFavors", source = "dailyFavors")
+    StatisticsDto.DailyFavorsResponse toDailyFavorsResponse(StaticInfo.DailyFavorsInfo info);
+
+    StatisticsDto.DailyFavorDto toDailyFavorDto(StaticInfo.DailyTravelFavorDetail detail);
 }
