+++ backend/yigil-admin/src/test/java/kr/co/yigil/stats/application/StatsFacadeTest.java
@@ -1,13 +1,7 @@
 package kr.co.yigil.stats.application;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
-
-import java.time.LocalDate;
-import java.util.Collections;
-import java.util.List;
 import kr.co.yigil.region.domain.DailyRegion;
+import kr.co.yigil.stats.domain.StaticInfo;
 import kr.co.yigil.stats.domain.StatsService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,6 +9,16 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.time.LocalDate;
+import java.util.Collections;
+import java.util.List;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class StatsFacadeTest {
@@ -36,5 +40,27 @@ class StatsFacadeTest {
         List<DailyRegion> actualRegionStats = statsFacade.getRegionStats(startDate, endDate);
 
         assertEquals(expectedRegionStats, actualRegionStats);
+    }
+
+
+    @DisplayName("일별 좋아요 수를 조회한다.")
+    @Test
+    void getDailyFavors() {
+        when(statsService.getDailyFavors(any(), any(), any(), any())).thenReturn(mock(StaticInfo.DailyFavorsInfo.class));
+
+        var result = statsFacade.getDailyFavors(null, null, null, null);
+
+        assertThat(result).isInstanceOf(StaticInfo.DailyFavorsInfo.class);
+    }
+
+
+    @DisplayName("일별 좋아요 수 상위를 조회한다.")
+    @Test
+    void getTopDailyFavors() {
+        when(statsService.getTopDailyFavors(any(), any(), any(), any())).thenReturn(mock(StaticInfo.DailyFavorsInfo.class));
+
+        var result = statsFacade.getTopDailyFavors(null, null, null, null);
+
+        assertThat(result).isInstanceOf(StaticInfo.DailyFavorsInfo.class);
     }
 }
