+++ backend/yigil-admin/src/test/java/kr/co/yigil/stats/infrastructure/StatsReaderImplTest.java
@@ -1,20 +1,27 @@
 package kr.co.yigil.stats.infrastructure;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
-
-import java.time.LocalDate;
-import java.util.Collections;
-import java.util.List;
+import kr.co.yigil.favor.domain.DailyFavorCount;
+import kr.co.yigil.favor.infrastructure.DailyFavorCountRepository;
 import kr.co.yigil.region.domain.DailyRegion;
 import kr.co.yigil.region.infrastructure.DailyRegionRepository;
+import kr.co.yigil.travel.domain.Travel;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.data.domain.PageImpl;
+
+import java.time.LocalDate;
+import java.util.Collections;
+import java.util.List;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 
 @ExtendWith(MockitoExtension.class)
@@ -22,9 +29,10 @@ class StatsReaderImplTest {
 
     @InjectMocks
     private StatsReaderImpl statsReader;
-
     @Mock
     private DailyRegionRepository dailyRegionRepository;
+    @Mock
+    private DailyFavorCountRepository dailyFavorCountRepository;
 
     @DisplayName("getRegionStats 메서드가 DailyRegionRepository를 잘 호출하는지")
     @Test
@@ -37,5 +45,29 @@ class StatsReaderImplTest {
         List<DailyRegion> actualRegionStats = statsReader.getRegionStats(startDate, endDate);
 
         assertEquals(expectedRegionStats, actualRegionStats);
+    }
+
+    @DisplayName("특정 기간 동안의 일별 좋아요 수 조회가 잘 되는지 확인한다.")
+    @Test
+    void readDailyFavorCountBetween() {
+
+        DailyFavorCount dailyFavorCount = new DailyFavorCount(5L, mock(Travel.class), null, null);
+        when(dailyFavorCountRepository.findAllByCreatedAtBetweenAndTravelTypeOrAllOrderByCountDesc(any(), any(), any(), any())).thenReturn(new PageImpl<>(List.of(dailyFavorCount)));
+
+        var result = statsReader.readDailyFavorCountBetween(null, null, null, null);
+
+        assertThat(result).isInstanceOf(PageImpl.class);
+    }
+
+
+    @DisplayName("특정 기간 동안의 상위 N개의 일별 좋아요 수 조회가 잘 되는지 확인한다.")
+    @Test
+    void getTopDailyFavorCount() {
+        DailyFavorCount dailyFavorCount = new DailyFavorCount(5L, mock(Travel.class), null, null);
+        when(dailyFavorCountRepository.findTopByTravelTypeOrderByCountDesc(any(), any(), any(), any())).thenReturn(List.of(dailyFavorCount));
+
+        var result = statsReader.getTopDailyFavorCount(null, null, null, null);
+
+        assertThat(result).isInstanceOf(List.class);
     }
 }
